+++ src/common/helpers/navigation.helper.ts
@@ -25,3 +25,31 @@ export const getResourceIdFromUri = () => {
 
   return splittedPath[resourceIndex + 1];
 };
+
+export const generatePageURL = ({
+  url,
+  queryParams,
+  profileId,
+}: {
+  url: string;
+  queryParams: Record<QueryParams, string>;
+  profileId: string;
+}) => {
+  const urlParams = new URLSearchParams();
+
+  if (queryParams) {
+    Object.entries(queryParams).forEach(([key, value]) => {
+      if (value) {
+        urlParams.append(key, value);
+      }
+    });
+  }
+
+  if (profileId) {
+    urlParams.set(QueryParams.ProfileId, profileId);
+  }
+
+  const paramString = urlParams.toString();
+
+  return paramString ? `${url}?${paramString}` : url;
+};