+++ src/components/ModalChooseProfile/index.ts
@@ -1,0 +1,1 @@
+export { ModalChooseProfile } from './ModalChooseProfile';