+++ src/store/selectors.ts
@@ -7,6 +7,7 @@ import { useConfigStore } from './stores/config';
 import { useUIStore } from './stores/ui';
 import { useSearchStore } from './stores/search';
 import { useComplexLookupStore } from './stores/complexLookup';
+import { useNavigationStore } from './stores/navigation';
 
 // The "createSelectors" function can be utilized here to generate memoized selectors
 // Note: "createSelectors" is currently unoptimized and may result in memory leaks if used as is
@@ -19,3 +20,4 @@ export const useConfigState = useConfigStore;
 export const useUIState = useUIStore;
 export const useSearchState = useSearchStore;
 export const useComplexLookupState = useComplexLookupStore;
+export const useNavigationState = useNavigationStore;