+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Address accessibility issues. Refs [UILD-579].
 * Add support for serials. Refs [UILD-594].
 * Add support for updated extent handling. Rfs [UILD-557].
+* Added Profile selection when the user creates a new Instance. Refs[UILD-573].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -37,6 +38,7 @@
 [UILD-579]:https://folio-org.atlassian.net/browse/UILD-579
 [UILD-594]:https://folio-org.atlassian.net/browse/UILD-594
 [UILD-557]:https://folio-org.atlassian.net/browse/UILD-557
+[UILD-573]:https://folio-org.atlassian.net/browse/UILD-573
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].